# swagger.json
{
  "swagger": "2.0.0",
  "info": {
    "version": "1.0.0",
    "title": "ParkLib's API",
    "description": "Parking location API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization:Bearer",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Authorization: Bearer"
    }
  },
  "basePath": "/",
...
  "paths": {
...
    "/auth/": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User that we want to login",
            "schema": {
              "$ref": "#/definitions/Login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ResponseAuth"
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": ["Users"],
        "summary": "Get all users",
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ResponseUsers"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "properties": {
        "id": {
          "type": "integer",
          "example": 1
        },
        "firstName": {
          "type": "string",
          "example": "John"
        },
        "lastName": {
          "type": "string",
          "example": "Doe"
        },
        "email": {
          "type": "string",
            "example": "johndoe@example.com"
        },
        "address": {
          "type": "string",
          "example": "1234 Main St"
        },
        "phone": {
          "type": "string",
          "example": "123-456-7890"
        },
        "picture": {
          "type": "string",
          "example": "https://example.com/picture.jpg"
        }
      }
    },
...
    "Login": {
      "required": ["email","password"],
      "properties": {
        "email": {
          "type": "string",
          "example": "johndoe@example.com"
        },
        "password": {
          "type": "string",
          "example": "password"
        }
      }
    },
    "ResponseAuth": {
      "properties": {
        "timeStamp": {
          "type": "object",
          "description": "Timestamp of the response",
          "example": "11/17/2022"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "httpStatus": {
          "type": "string",
          "example": "OK"
        },
        "message": {
          "type": "string",
          "example": "Success"
        },
        "data": {
          "properties": {
            "accesToken": {
              "type": "string",
              "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.="
            }
          }
        }
      }
    },
    "ResponseUsers": {
      "properties": {
        "timeStamp": {
          "type": "object",
          "description": "Timestamp of the response",
          "example": "11/17/2022"
        },
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "httpStatus": {
          "type": "string",
          "example": "OK"
        },
        "message": {
          "type": "string",
          "example": "Success"
        },
        "data": {
          "type": "array",
          "$ref": "#/definitions/User"
        }
      }
    },
...
  },
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"]
}